miniapp/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
